+++ CHANGELOG.md
@@ -31,6 +31,7 @@
  * **@dlr-eoc/map-maplibre:**
  - Allow the usage of different coordinate reference systems in WMTS requests. Use `WmtsLayer.params.matrixSetOptions.matrixSet`. and `WmtsLayer.params.matrixSetOptions.tileMatrixPrefix`
  - Add more types and fix type errors
+ - Fix KMLSource `setData` add `waitForCompletion`
 
 
 # [16.0.0](https://github.com/dlr-eoc/ukis-frontend-libraries/tree/v16.0.0) (2026-05-29) (update dependencies, new library ngx-ukis-ui-clarity and ngx-ukis-utilities, replace core-ui)